Tests/LibWeb/Text/expected/HTML/dimension-attributes.txt

Test hr.width = "100" maps to width: 100px
Test hr.width = " 00110 " maps to width: 110px
Test hr.width = "120." maps to width: 120px
